services/ingestion/requirements.txt
fastapi
uvicorn[standard]
aiokafka
kafka-python